wetwet

+++ Userland/SampleCodeModule/include/stdlib.h
@@ -38,7 +38,7 @@ void srand(unsigned long int seed);
 void lcg(unsigned long int *x, unsigned long int a, int c,
          unsigned long int m) ;
 
-void* memcpy(void* destination, const void* source, size_t length);
+void* memcpy(void* destination, const void* source, unsigned long long length);
         
 
 unsigned long int rand();
